graphic and UI developement, added weird ending sequence

+++ gameModel.h
@@ -13,7 +13,8 @@ class gameModel
         void increasePoints(int);
         int getPoints();
     private:
-        int points;
+        long points;
+
 };
